ilri_ges/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-filter==22.1
django-simple-history==3.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.39.0
gunicorn==20.1.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.7.1
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
xlrd==2.0.1
zipp==3.14.0
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-filter==22.1
django-simple-history==3.2.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.39.0
gunicorn==20.1.0
importlib-metadata==6.0.0
importlib-resources==5.12.0
kiwisolver==1.4.4
Markdown==3.4.1
matplotlib==3.7.1
numpy==1.24.2
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pycodestyle==2.10.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
xlrd==2.0.1
zipp==3.14.0
